+++ src/frequenz/sdk/timeseries/__init__.py
@@ -1,0 +1,26 @@
+"""
+Handling of timeseries streams.
+
+A timeseries is a stream (normally an async iterator) of
+[samples][frequenz.sdk.timeseries.Sample].
+
+This module provides tools to operate on timeseries.
+
+Copyright
+Copyright © 2022 Frequenz Energy-as-a-Service GmbH
+
+License
+MIT
+"""
+
+from .resampler import (
+    ComponentMetricGroupResampler,
+    ComponentMetricResampler,
+    ResamplingFunction,
+)
+
+__all__ = [
+    "ComponentMetricGroupResampler",
+    "ComponentMetricResampler",
+    "ResamplingFunction",
+]
